+++ server/products/forms/categories.py
@@ -1,5 +1,5 @@
 from django import forms
-from .models import Category
+from products.models import Category
 
 class CategoryForm(forms.Form):
     name = forms.CharField(max_length=255,
