+++ server/products/templates/categories/delete.html
@@ -1,0 +1,13 @@
+{% extends 'server/master.html' %}
+
+{% block page_title %}
+Delete item
+{% endblock %}
+
+{% block page_content %}
+<form method="post">
+    {% csrf_token %}
+    <span> Item {{ obj.name }} delete.</span>
+    <input type="submit">
+</form>
+{% endblock %}
