+++ server/products/templates/categories/update.html
@@ -1,7 +1,7 @@
 {% extends 'server/master.html' %}
 
 {% block page_title %}
-Detail products
+Update item
 {% endblock %}
 
 {% block page_content %}
