+++ server/server/templates/server/form.html
@@ -1,0 +1,21 @@
+<form method = "POST" enctype="multipart/form-data">
+    {% for err in form.not_field_errors %}
+        <div clas="alert alert-danger">
+            <span>{{ err }}</span>
+        </div>
+    {% endfor %}
+
+    {% csrf_token %}
+    {% for field in form %}
+    <div clas="field-group">
+        <label for="{{ field.name }}">{{ field.label }}</label>
+        {{ field }}
+
+        {% for err in field.errors %}
+            <small class="fild_error">{{ err }}</small>
+        {% endfor %}
+
+    </div>
+    {% endfor %}
+    <input type="submit">
+</form>
